+++ dll/win32/shell32/shelldesktop/CChangeNotify.h
