+++ dll/win32/shell32/shelldesktop/shelldesktop.h
@@ -29,6 +29,7 @@
 #include <undocuser.h>
 
 #include <shellutils.h>
+#include "CChangeNotify.h"
 #include "../shresdef.h"
 
 #include <wine/debug.h>
